hydrate polyfuse-mount into polyfuse
