--- crates/polyfuse-mount/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "polyfuse-mount"
-version = "0.1.0-dev"
-publish = false
-authors = ["Yusuke Sasaki <yusuke.sasaki.nuem@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-libc = "0.2"
