+++ crates/polyfuse/src/lib.rs
@@ -1,10 +1,15 @@
 #![warn(clippy::todo, clippy::unimplemented)]
 
+#[macro_use]
+mod syscall;
+
 mod decoder;
+mod mount;
 mod session;
 mod write;
 
 pub mod bytes;
+pub mod conn;
 pub mod op;
 pub mod reply;
